+++ CHANGELOG.md
@@ -75,10 +75,10 @@
   * `RecipientNotFound`
   * `UserForbidsGroups`
   * `UserKicked`
-* Added `EmbeddedGallery` and `EmbeddedGalleryImage` types to support the new [image gallery format](https://github.com/bluesky-social/atproto/pull/4827).
-  Note that `EmbeddedGalleryImage` requires an aspect ratio, unlike the `EmbeddedImage` type.
+* Added `Embed.EmbeddedGallery` and `Embed.Gallery.GalleryImage` types to support the new [image gallery format](https://github.com/bluesky-social/atproto/pull/4827).
+  Note that `GalleryImage` requires an aspect ratio, unlike the `EmbeddedImage` type.
 * Added new overloads to `BlueskyAgent.Post` to support `EmbeddedGallery` content.
-  If you pass more than 4 `EmbeddedImage` items to `Post()`, it will attempt to convert them to `EmbeddedGallery` format. Your `EmbeddedImage`
+  If you pass more than 4 `EmbeddedImage` items to `Post()`, it will attempt to convert them an `EmbeddedGallery`. Your `EmbeddedImage`
   instances must have an aspect ratio for this to work, otherwise an exception will be thrown.
   You can use [Magick.NET](https://github.com/dlemstra/Magick.NET), [ImageSharp](https://github.com/SixLabors/ImageSharp) or other image processing libraries to calculate the aspect ratio of your images if needed.
 * Added new overloads to `PostBuilder` to support `EmbeddedGallery` content.